pyproject.toml
[tool.poetry]
name = "bbdc"
version = "0.1.0"
description = ""
authors = ["remykarem <brushed-pascal.09@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
pydantic = "^2.1.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
